Add files via upload
I think I managed to create 1000 exitatory and 250 inhibitory connection for every neuron in the function connect of network.cpp. Moreover, i managed to create the function update of the class network.
I still have to work on the poisson distribution because it does not compile.
The next step is to make tests to now if i coded the right way.